Combine AddService and EditService to ServiceForm Component
