+++ src/app/app.routes.ts
@@ -25,8 +25,7 @@ import { RoomsComponent } from './components/shared/rooms/rooms.component';
 import { RoomFormComponent } from './components/admin/room-form/room-form.component';
 import { RoomDetailsComponent } from './components/shared/room-details/room-details.component';
 import { ServicesComponent } from './components/employee/services/services.component';
-import { AddServiceComponent } from './components/employee/add-service/add-service.component';
-import { EditServiceComponent } from './components/employee/edit-service/edit-service.component';
+import { ServiceFormComponent } from './components/employee/service-form/service-form.component';
 
 export const authGuard: CanActivateFn = (
   route: ActivatedRouteSnapshot,
@@ -92,13 +91,13 @@ export const routes: Routes = [
   },
   {
     path: 'employee/add-service',
-    component: AddServiceComponent,
+    component: ServiceFormComponent,
     canActivate: [authGuard],
     data: { roles: ['Employee'] },
   },
   {
     path: 'employee/edit-service/:id',
-    component: EditServiceComponent,
+    component: ServiceFormComponent,
     canActivate: [authGuard],
     data: { roles: ['Employee'] },
   },
