--- src/app/components/employee/edit-service/edit-service.component.scss
@@ -1,149 +1,0 @@
-@import "../../../../styles/variables.scss";
-
-.page {
-  max-width: 900px;
-  width: auto;
-  margin-left: auto;
-  margin-right: auto;
-}
-
-.center-wrapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.container {
-  background-color: white;
-  margin: auto;
-  padding: 15px;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  width: 100%;
-  border-radius: 10px;
-  max-width: 900px;
-}
-
-.input-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.input-error {
-  width: 50%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-label {
-  width: 25%;
-}
-
-.input-group-icon {
-  width: 100%;
-  margin: 10px 5px;
-  padding: 11px;
-  font-size: 16px;
-  border: 2px $color-dark-brown solid;
-  border-right: none;
-}
-
-input,
-select {
-  margin: 10px 5px;
-  padding: 7px;
-  font-size: 16px;
-  border: 2px $color-dark-brown solid;
-  border-radius: 5px;
-  width: 100%;
-  border-left: none;
-}
-
-.type-input {
-  border-right: none;
-  margin-right: 0px;
-}
-
-.right-input .input-group-icon {
-  border: 2px $color-dark-brown solid;
-  border-left: none;
-  border-radius: 0 5px 5px 0;
-}
-
-.right-input input {
-  border: 2px $color-dark-brown solid;
-  border-right: none;
-  border-radius: 5px 0 0 5px;
-}
-
-.right-input .radio-container {
-  border: 2px $color-dark-brown solid;
-  border-right: none;
-  border-radius: 5px 0 0 5px;
-}
-
-.form-control:focus {
-  box-shadow: none !important;
-  outline: none !important;
-  border: 2px $color-dark-brown solid;
-  border-left: none;
-}
-
-.type-input:focus {
-  border-right: none;
-}
-
-.error {
-  color: $color-red;
-  text-align: start;
-  width: 100%;
-  font-size: 14px;
-  margin: -6px 10px 10px 50px;
-  font-weight: bold;
-}
-
-button {
-  text-decoration: none;
-  width: auto;
-  margin: auto;
-  font-size: 15px;
-  border-radius: 7px;
-  padding: 7px 10px;
-}
-
-.btn-submit {
-  border: none;
-  font-size: 18px;
-  background-color: $color-dark-brown;
-  color: white;
-}
-
-i {
-  color: $color-dark-brown;
-}
-
-.input-group-right-icons {
-  width: 100%;
-  font-size: 16px;
-  border: 2px $color-dark-brown solid;
-  border-left: none;
-  border-radius: 0 5px 5px 0;
-  margin-right: 5px;
-  padding: 8px 12px 9.5px 12px;
-}
-
-.right-input .input-group-right-icons {
-  border: 2px $color-dark-brown solid;
-  border-left: none;
-  border-radius: 5px 0 0 5px;
-}
-
-input::-webkit-outer-spin-button,
-input::-webkit-inner-spin-button {
-  -webkit-appearance: none;
-  margin: 0;
-}
